pyproject.toml
[tool.poetry]
name = "grpc-microservices"
version = "0.1.0"
description = ""
authors = ["frankdelan <frank.danil704@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
grpcio = "^1.64.1"
grpcio-tools = "^1.64.1"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
protobuf = "^5.27.2"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
